requirements.txt
streamlit
pandas
numpy
sqlalchemy
kaggle
pymysql
patool
cryptography
plotly.express
mysql-connector-python
pytest